refactor: move token related code to a common dir
