+++ packages/frontend/src/components/common/token/TokenIcon.js
